POST /members/reauth
JWT are short-lived (in fact, we're also cutting it here down to 15m instead of 30m), so we have POST /members/reauth to issue new ones. Clients using the API should be advised to check how old the token is each time it's used, and if it's more than 10 minutes old, get a new one.