Better organize provider endpoints
There's a principle in REST API design that says that your endpoints should be nouns, and then you can rely on verbs to decide what to do with those nouns. I don't know if that makes sense throughout the entire API that we have here, but it certainly makes sense here. This organizes the endpoints that deal with OAuth 2.0 tokens under a /members/providers endpoint, with a POST to create a new one, a GET to fetch the array, and a DELETE /members/providers/:provider to delete a specific one.