Update documentation on Page objects
Documenting the Page object highlighted how we should clean it up, so we introduced the Page.export() method to do that, and set up all of the Pages API endpoints to use it. This commit brings us full circle, updating the documentation to reflect the change that writing the documentation inspired.