POST /pages - Return 400 when you try to create page with existing path
Paths need to be unique. If you try to create a new page that has a path that is the same as some other page's path, the API endpoint returns a 400 ("Bad Request") error. Since this is the only reason why a page might be rejected, when you get a 400 error, you know it's because there's already a page with the given path.