TemplateHandler.query - Fix SQL
The bug in the old SQL was that if there were any pages with the template, it would return all the pages. So if there were no pages with the template, tests would pass. This commit fixes the SQL and updates the two unit tests to check for that condition.