The Mother of All MySQL Queries
Well this got complicated. In order for our limits, offsets, and other filters to work, we need to get everything in our initial query. That means pulling that tag logic in. And that gets complicated, because with an OR, that just means adding a join and some new conditions, but with an AND, that means adding a series of joins, one for each tag conditional. This makes our already most-complicated, core SQL query even more complicated, but it also makes it amazingly powerful.